refactor: extract ThreadSidebar to its own file for reuse in app sidebar
